requirements.txt
# FastAPI and web server
fastapi>=0.95.0
uvicorn>=0.22.0
gunicorn>=20.1.0
python-multipart>=0.0.6

# Celery and Redis
celery>=5.3.0
redis>=4.5.5

# Cohere API
cohere>=4.5.0

# Environment variables and logging
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
httpx>=0.24.0
tenacity>=8.2.2

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
